vpratt

Trait VprattCore

pub trait VprattCore {
    type Item;
    type PrattToken: Clone + Copy + PartialEq;
    type Output;
    type Error;
}

The internal trait powering the Left Binding Power (LBP) loop.

This trait is automatically implemented for your parser struct by the #[parser] attribute macro. It contains the hidden state mechanisms required to drive Pratt parsing without allocating.

Required Associated Types

type Item

The raw token type yielded by the supplied stream iterator.

type PrattToken: Clone + Copy + PartialEq

The simplified routing token (usually a TokenKind enum) extracted from the Item.

type Output

The Abstract Syntax Tree (AST) node type returned.

type Error

The error type returned by the handlers.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.